requirements.txt
beautifulsoup4
requests
numpy
Pillow

